requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.0
python-dotenv==1.0.1
scipy>=1.10.0
numpy>=1.24.0
tensorflow>=2.14.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
